Quick fix for bug #79
Maybe not the most ideal way to fix this, but it does prevent the
'single' checkbox type from subscribing people who didn't opt in. Does
not affect the other types, although the dropdown select never did work
due to a different bug. A better way to fix this might be to add
'ctctcf7_lists' to the 'Opt-In Field' dropdown list on the Constant
Contact tab, but I haven't yet found where to add that - maybe something
the plugin author can address more adequately.